@@ -10,9 +10,10 @@
         }
     ],
     "require": {
-        "php": ">=5.5.0",
-        "guzzlehttp/guzzle": "~6.0",
-        "ext-mbstring": "*"
+        "php": ">=7.4",
+        "guzzlehttp/guzzle": "7.*",
+        "ext-mbstring": "*",
+        "ext-json": "*"
     },
     "require-dev": {
         "phpunit/phpunit": "4.2.*",
