@@ -2,133 +2,128 @@
 
 namespace Maknz\Slack;
 
-class AttachmentField
-{
-    /**
-     * The required title field of the field.
-     *
-     * @var string
-     */
-    protected $title;
+class AttachmentField {
+	/**
+	 * The required title field of the field.
+	 *
+	 * @var string
+	 */
+	protected string $title;
 
-    /**
-     * The required value of the field.
-     *
-     * @var string
-     */
-    protected $value;
+	/**
+	 * The required value of the field.
+	 *
+	 * @var string
+	 */
+	protected string $value;
 
-    /**
-     * Whether the value is short enough to fit side by side with
-     * other values.
-     *
-     * @var bool
-     */
-    protected $short = false;
+	/**
+	 * Whether the value is short enough to fit side by side with
+	 * other values.
+	 *
+	 * @var bool
+	 */
+	protected bool $short = false;
 
-    /**
-     * Instantiate a new AttachmentField.
-     *
-     * @param array $attributes
-     * @return void
-     */
-    public function __construct(array $attributes)
-    {
-        if (isset($attributes['title'])) {
-            $this->setTitle($attributes['title']);
-        }
+	/**
+	 * Instantiate a new AttachmentField.
+	 *
+	 * @param array $attributes
+	 *
+	 * @return void
+	 */
+	public function __construct(array $attributes) {
+		if (isset($attributes['title'])) {
+			$this->setTitle($attributes['title']);
+		}
 
-        if (isset($attributes['value'])) {
-            $this->setValue($attributes['value']);
-        }
+		if (isset($attributes['value'])) {
+			$this->setValue($attributes['value']);
+		}
 
-        if (isset($attributes['short'])) {
-            $this->setShort($attributes['short']);
-        }
-    }
+		if (isset($attributes['short'])) {
+			$this->setShort($attributes['short']);
+		}
+	}
 
-    /**
-     * Get the title of the field.
-     *
-     * @return string
-     */
-    public function getTitle()
-    {
-        return $this->title;
-    }
+	/**
+	 * Get the title of the field.
+	 *
+	 * @return string
+	 */
+	public function getTitle(): string {
+		return $this->title;
+	}
 
-    /**
-     * Set the title of the field.
-     *
-     * @param string $title
-     * @return $this
-     */
-    public function setTitle($title)
-    {
-        $this->title = $title;
+	/**
+	 * Set the title of the field.
+	 *
+	 * @param string $title
+	 *
+	 * @return $this
+	 */
+	public function setTitle(string $title): self {
+		$this->title = $title;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the value of the field.
-     *
-     * @return string
-     */
-    public function getValue()
-    {
-        return $this->value;
-    }
+	/**
+	 * Get the value of the field.
+	 *
+	 * @return string
+	 */
+	public function getValue(): string {
+		return $this->value;
+	}
 
-    /**
-     * Set the value of the field.
-     *
-     * @param string $value
-     * @return $this
-     */
-    public function setValue($value)
-    {
-        $this->value = $value;
+	/**
+	 * Set the value of the field.
+	 *
+	 * @param string $value
+	 *
+	 * @return $this
+	 */
+	public function setValue(string $value): self {
+		$this->value = $value;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get whether this field is short enough for displaying
-     * side-by-side with other fields.
-     *
-     * @return bool
-     */
-    public function getShort()
-    {
-        return $this->short;
-    }
+	/**
+	 * Get whether this field is short enough for displaying
+	 * side-by-side with other fields.
+	 *
+	 * @return bool
+	 */
+	public function getShort(): bool {
+		return $this->short;
+	}
 
-    /**
-     * Set whether this field is short enough for displaying
-     * side-by-side with other fields.
-     *
-     * @param string $value
-     * @return $this
-     */
-    public function setShort($value)
-    {
-        $this->short = (bool) $value;
+	/**
+	 * Set whether this field is short enough for displaying
+	 * side-by-side with other fields.
+	 *
+	 * @param string $value
+	 *
+	 * @return $this
+	 */
+	public function setShort(string $value): self {
+		$this->short = (bool)$value;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the array representation of this attachment field.
-     *
-     * @return array
-     */
-    public function toArray()
-    {
-        return [
-            'title' => $this->getTitle(),
-            'value' => $this->getValue(),
-            'short' => $this->getShort(),
-        ];
-    }
+	/**
+	 * Get the array representation of this attachment field.
+	 *
+	 * @return array
+	 */
+	public function toArray(): array {
+		return [
+			'title' => $this->getTitle(),
+			'value' => $this->getValue(),
+			'short' => $this->getShort(),
+		];
+	}
 }
