@@ -4,402 +4,393 @@ namespace Maknz\Slack;
 
 use InvalidArgumentException;
 
-class Message
-{
-    /**
-     * Reference to the Slack client responsible for sending
-     * the message.
-     *
-     * @var \Maknz\Slack\Client
-     */
-    protected $client;
+class Message {
+	/**
+	 * Reference to the Slack client responsible for sending
+	 * the message.
+	 *
+	 * @var Client
+	 */
+	protected Client $client;
 
-    /**
-     * The text to send with the message.
-     *
-     * @var string
-     */
-    protected $text;
+	/**
+	 * The text to send with the message.
+	 *
+	 * @var string
+	 */
+	protected string $text;
 
-    /**
-     * The channel the message should be sent to.
-     *
-     * @var string
-     */
-    protected $channel;
+	/**
+	 * The channel the message should be sent to.
+	 *
+	 * @var string
+	 */
+	protected string $channel;
 
-    /**
-     * The username the message should be sent as.
-     *
-     * @var string
-     */
-    protected $username;
+	/**
+	 * The username the message should be sent as.
+	 *
+	 * @var string
+	 */
+	protected string $username;
 
-    /**
-     * The URL to the icon to use.
-     *
-     * @var string
-     */
-    protected $icon;
+	/**
+	 * The URL to the icon to use.
+	 *
+	 * @var string
+	 */
+	protected string $icon;
 
-    /**
-     * The type of icon we are using.
-     *
-     * @var enum
-     */
-    protected $iconType;
+	/**
+	 * The type of icon we are using.
+	 *
+	 * @var string
+	 */
+	protected string $iconType;
 
-    /**
-     * Whether the message text should be interpreted in Slack's
-     * Markdown-like language.
-     *
-     * @var bool
-     */
-    protected $allow_markdown = true;
+	/**
+	 * Whether the message text should be interpreted in Slack's
+	 * Markdown-like language.
+	 *
+	 * @var bool
+	 */
+	protected bool $allow_markdown = true;
 
-    /**
-     * The attachment fields which should be formatted with
-     * Slack's Markdown-like language.
-     *
-     * @var array
-     */
-    protected $markdown_in_attachments = [];
+	/**
+	 * The attachment fields which should be formatted with
+	 * Slack's Markdown-like language.
+	 *
+	 * @var array
+	 */
+	protected array $markdown_in_attachments = [];
 
-    /**
-     * An array of attachments to send.
-     *
-     * @var array
-     */
-    protected $attachments = [];
+	/**
+	 * An array of attachments to send.
+	 *
+	 * @var array
+	 */
+	protected array $attachments = [];
 
-    /**
-     * @var string
-     */
-    const ICON_TYPE_URL = 'icon_url';
+	/**
+	 * @var string
+	 */
+	public const ICON_TYPE_URL = 'icon_url';
 
-    /**
-     * @var string
-     */
-    const ICON_TYPE_EMOJI = 'icon_emoji';
+	/**
+	 * @var string
+	 */
+	public const ICON_TYPE_EMOJI = 'icon_emoji';
 
-    /**
-     * Instantiate a new Message.
-     *
-     * @param \Maknz\Slack\Client $client
-     * @return void
-     */
-    public function __construct(Client $client)
-    {
-        $this->client = $client;
-    }
+	/**
+	 * Instantiate a new Message.
+	 *
+	 * @param Client $client
+	 *
+	 * @return void
+	 */
+	public function __construct(Client $client) {
+		$this->client = $client;
+	}
 
-    /**
-     * Get the message text.
-     *
-     * @return string
-     */
-    public function getText()
-    {
-        return $this->text;
-    }
+	/**
+	 * Get the message text.
+	 *
+	 * @return string
+	 */
+	public function getText(): string {
+		return $this->text;
+	}
 
-    /**
-     * Set the message text.
-     *
-     * @param string $text
-     * @return $this
-     */
-    public function setText($text)
-    {
-        $this->text = $text;
+	/**
+	 * Set the message text.
+	 *
+	 * @param string $text
+	 *
+	 * @return $this
+	 */
+	public function setText(string $text): self {
+		$this->text = $text;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the channel we will post to.
-     *
-     * @return string
-     */
-    public function getChannel()
-    {
-        return $this->channel;
-    }
+	/**
+	 * Get the channel we will post to.
+	 *
+	 * @return string
+	 */
+	public function getChannel(): string {
+		return $this->channel;
+	}
 
-    /**
-     * Set the channel we will post to.
-     *
-     * @param string $channel
-     * @return $this
-     */
-    public function setChannel($channel)
-    {
-        $this->channel = $channel;
+	/**
+	 * Set the channel we will post to.
+	 *
+	 * @param string $channel
+	 *
+	 * @return $this
+	 */
+	public function setChannel(string $channel): self {
+		$this->channel = $channel;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the username we will post as.
-     *
-     * @return string
-     */
-    public function getUsername()
-    {
-        return $this->username;
-    }
+	/**
+	 * Get the username we will post as.
+	 *
+	 * @return string
+	 */
+	public function getUsername(): string {
+		return $this->username;
+	}
 
-    /**
-     * Set the username we will post as.
-     *
-     * @param string $username
-     * @return $this
-     */
-    public function setUsername($username)
-    {
-        $this->username = $username;
+	/**
+	 * Set the username we will post as.
+	 *
+	 * @param string $username
+	 *
+	 * @return $this
+	 */
+	public function setUsername(string $username): self {
+		$this->username = $username;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the icon (either URL or emoji) we will post as.
-     *
-     * @return string
-     */
-    public function getIcon()
-    {
-        return $this->icon;
-    }
+	/**
+	 * Get the icon (either URL or emoji) we will post as.
+	 *
+	 * @return string
+	 */
+	public function getIcon(): string {
+		return $this->icon;
+	}
 
-    /**
-     * Set the icon (either URL or emoji) we will post as.
-     *
-     * @param string $icon
-     * @return $this
-     */
-    public function setIcon($icon)
-    {
-        if ($icon == null) {
-            $this->icon = $this->iconType = null;
+	/**
+	 * Set the icon (either URL or emoji) we will post as.
+	 *
+	 * @param string $icon
+	 *
+	 * @return Message|void
+	 */
+	public function setIcon(string $icon) {
+		if ($icon === null) {
+			/** @noinspection PhpFieldAssignmentTypeMismatchInspection */
+			$this->icon = $this->iconType = null;
 
-            return;
-        }
+			return;
+		}
 
-        if (mb_substr($icon, 0, 1) == ':' && mb_substr($icon, mb_strlen($icon) - 1, 1) == ':') {
-            $this->iconType = self::ICON_TYPE_EMOJI;
-        } else {
-            $this->iconType = self::ICON_TYPE_URL;
-        }
+		if (mb_strpos($icon, ':') === 0 && mb_substr($icon, mb_strlen($icon) - 1, 1) === ':') {
+			$this->iconType = self::ICON_TYPE_EMOJI;
+		} else {
+			$this->iconType = self::ICON_TYPE_URL;
+		}
 
-        $this->icon = $icon;
+		$this->icon = $icon;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the icon type being used, if an icon is set.
-     *
-     * @return string
-     */
-    public function getIconType()
-    {
-        return $this->iconType;
-    }
+	/**
+	 * Get the icon type being used, if an icon is set.
+	 *
+	 * @return string
+	 */
+	public function getIconType(): string {
+		return $this->iconType;
+	}
 
-    /**
-     * Get whether message text should be formatted with
-     * Slack's Markdown-like language.
-     *
-     * @return bool
-     */
-    public function getAllowMarkdown()
-    {
-        return $this->allow_markdown;
-    }
+	/**
+	 * Get whether message text should be formatted with
+	 * Slack's Markdown-like language.
+	 *
+	 * @return bool
+	 */
+	public function getAllowMarkdown(): bool {
+		return $this->allow_markdown;
+	}
 
-    /**
-     * Set whether message text should be formatted with
-     * Slack's Markdown-like language.
-     *
-     * @param bool $value
-     * @return void
-     */
-    public function setAllowMarkdown($value)
-    {
-        $this->allow_markdown = (bool) $value;
+	/**
+	 * Set whether message text should be formatted with
+	 * Slack's Markdown-like language.
+	 *
+	 * @param bool $value
+	 *
+	 * @return Message
+	 */
+	public function setAllowMarkdown(bool $value): Message {
+		$this->allow_markdown = $value;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Enable Markdown formatting for the message.
-     *
-     * @return void
-     */
-    public function enableMarkdown()
-    {
-        $this->setAllowMarkdown(true);
+	/**
+	 * Enable Markdown formatting for the message.
+	 *
+	 * @return Message
+	 */
+	public function enableMarkdown(): Message {
+		$this->setAllowMarkdown(true);
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Disable Markdown formatting for the message.
-     *
-     * @return void
-     */
-    public function disableMarkdown()
-    {
-        $this->setAllowMarkdown(false);
+	/**
+	 * Disable Markdown formatting for the message.
+	 *
+	 * @return Message
+	 */
+	public function disableMarkdown(): Message {
+		$this->setAllowMarkdown(false);
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Get the attachment fields which should be formatted
-     * in Slack's Markdown-like language.
-     *
-     * @return array
-     */
-    public function getMarkdownInAttachments()
-    {
-        return $this->markdown_in_attachments;
-    }
+	/**
+	 * Get the attachment fields which should be formatted
+	 * in Slack's Markdown-like language.
+	 *
+	 * @return array
+	 */
+	public function getMarkdownInAttachments(): array {
+		return $this->markdown_in_attachments;
+	}
 
-    /**
-     * Set the attachment fields which should be formatted
-     * in Slack's Markdown-like language.
-     *
-     * @param array $fields
-     * @return void
-     */
-    public function setMarkdownInAttachments(array $fields)
-    {
-        $this->markdown_in_attachments = $fields;
+	/**
+	 * Set the attachment fields which should be formatted
+	 * in Slack's Markdown-like language.
+	 *
+	 * @param array $fields
+	 *
+	 * @return Message
+	 */
+	public function setMarkdownInAttachments(array $fields): Message {
+		$this->markdown_in_attachments = $fields;
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Change the name of the user the post will be made as.
-     *
-     * @param string $username
-     * @return $this
-     */
-    public function from($username)
-    {
-        $this->setUsername($username);
+	/**
+	 * Change the name of the user the post will be made as.
+	 *
+	 * @param string $username
+	 *
+	 * @return $this
+	 */
+	public function from(string $username): self {
+		$this->setUsername($username);
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Change the channel the post will be made to.
-     *
-     * @param string $channel
-     * @return $this
-     */
-    public function to($channel)
-    {
-        $this->setChannel($channel);
+	/**
+	 * Change the channel the post will be made to.
+	 *
+	 * @param string $channel
+	 *
+	 * @return $this
+	 */
+	public function to(string $channel): self {
+		$this->setChannel($channel);
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Chainable method for setting the icon.
-     *
-     * @param string $icon
-     * @return $this
-     */
-    public function withIcon($icon)
-    {
-        $this->setIcon($icon);
+	/**
+	 * Chainable method for setting the icon.
+	 *
+	 * @param string $icon
+	 *
+	 * @return $this
+	 */
+	public function withIcon(string $icon): self {
+		$this->setIcon($icon);
 
-        return $this;
-    }
+		return $this;
+	}
 
-    /**
-     * Add an attachment to the message.
-     *
-     * @param mixed $attachment
-     * @return $this
-     */
-    public function attach($attachment)
-    {
-        if ($attachment instanceof Attachment) {
-            $this->attachments[] = $attachment;
+	/**
+	 * Add an attachment to the message.
+	 *
+	 * @param mixed $attachment
+	 *
+	 * @return $this
+	 */
+	public function attach($attachment): self {
+		if ($attachment instanceof Attachment) {
+			$this->attachments[] = $attachment;
 
-            return $this;
-        } elseif (is_array($attachment)) {
-            $attachmentObject = new Attachment($attachment);
+			return $this;
+		}
 
-            if (! isset($attachment['mrkdwn_in'])) {
-                $attachmentObject->setMarkdownFields($this->getMarkdownInAttachments());
-            }
+		if (is_array($attachment)) {
+			$attachmentObject = new Attachment($attachment);
 
-            $this->attachments[] = $attachmentObject;
+			if (!isset($attachment['mrkdwn_in'])) {
+				$attachmentObject->setMarkdownFields($this->getMarkdownInAttachments());
+			}
 
-            return $this;
-        }
+			$this->attachments[] = $attachmentObject;
 
-        throw new InvalidArgumentException('Attachment must be an instance of Maknz\\Slack\\Attachment or a keyed array');
-    }
+			return $this;
+		}
 
-    /**
-     * Get the attachments for the message.
-     *
-     * @return array
-     */
-    public function getAttachments()
-    {
-        return $this->attachments;
-    }
+		throw new InvalidArgumentException('Attachment must be an instance of Maknz\\Slack\\Attachment or a keyed array');
+	}
 
-    /**
-     * Set the attachments for the message.
-     *
-     * @param array $attachments
-     * @return $this
-     */
-    public function setAttachments(array $attachments)
-    {
-        $this->clearAttachments();
+	/**
+	 * Get the attachments for the message.
+	 *
+	 * @return array
+	 */
+	public function getAttachments(): array {
+		return $this->attachments;
+	}
 
-        foreach ($attachments as $attachment) {
-            $this->attach($attachment);
-        }
+	/**
+	 * Set the attachments for the message.
+	 *
+	 * @param array $attachments
+	 *
+	 * @return $this
+	 */
+	public function setAttachments(array $attachments): self {
+		$this->clearAttachments();
 
-        return $this;
-    }
+		foreach ($attachments as $attachment) {
+			$this->attach($attachment);
+		}
 
-    /**
-     * Remove all attachments for the message.
-     *
-     * @return $this
-     */
-    public function clearAttachments()
-    {
-        $this->attachments = [];
+		return $this;
+	}
 
-        return $this;
-    }
+	/**
+	 * Remove all attachments for the message.
+	 *
+	 * @return $this
+	 */
+	public function clearAttachments(): self {
+		$this->attachments = [];
 
-    /**
-     * Send the message.
-     *
-     * @param string $text The text to send
-     * @return void
-     */
-    public function send($text = null)
-    {
-        if ($text) {
-            $this->setText($text);
-        }
+		return $this;
+	}
 
-        $this->client->sendMessage($this);
+	/**
+	 * Send the message.
+	 *
+	 * @param null $text The text to send
+	 *
+	 * @return Message
+	 */
+	public function send($text = null): Message {
+		if ($text) {
+			$this->setText($text);
+		}
 
-        return $this;
-    }
+		$this->client->sendMessage($this);
+
+		return $this;
+	}
 }
